Seagull added along with powerups
I have added in the seagull to the game but I am making a slight change to how it will be used.
I am going to have the seagull simply fly across the top of the screen and drop either the shield or the heart power up (scripted to be random and will drop between 15-25 seconds).